Change ttsend and ttrecv to repeat send/recv calls.
The send and recv functions are not guaranteed to send or receive as many bytes
as you want. In order to reliably send or receive the required amount of data,
you must check their return values and call them one or more times. This patch
does that, and fixes the breakage that happens otherwise when sending non-tiny
units of data to and from a remote server.